requirements.txt
PyOpenGL
wheel
